# src/modules/promo/dto/promo-stat-out.dto.ts
import { createZodDto } from 'nestjs-zod';
import { z } from 'zod';
import { createPaginatedSchema } from '@/common/dto/paginated.dto';

export const PromoStatSchema = z.object({
  id: z.string(),
  code: z.string(),
  discount: z.coerce.number(),
  activeFrom: z.coerce
    .date()
    .transform((d) => d.toISOString())
    .nullish(),
  expiredAt: z.coerce
    .date()
    .transform((d) => d.toISOString())
    .nullish(),
  globalLimit: z.coerce.number(),
  userLimit: z.coerce.number(),
  inactiveAt: z.coerce
    .date()
    .transform((d) => d.toISOString())
    .nullish(),
  createdAt: z.coerce
    .date()
    .transform((d) => d.toISOString())
    .nullish(),
  usageCount: z.coerce.number(),
});

export const PaginatedPromoStatSchema = createPaginatedSchema(PromoStatSchema);

export default class PaginatedPromoStatOutDto extends createZodDto(
  PaginatedPromoStatSchema,
) {}